+++ entities/portfolio/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @brains/portfolio
 
+## 0.2.0-alpha.351
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/atproto-contracts@0.2.0-alpha.351
+  - @brains/content-formatters@0.2.0-alpha.351
+  - @brains/contracts@0.2.0-alpha.351
+  - @brains/media-page-composer@0.2.0-alpha.351
+  - @brains/ui-library@0.2.0-alpha.351
+  - @brains/utils@0.2.0-alpha.351
+  - @brains/plugins@0.2.0-alpha.351
+  - @brains/templates@0.2.0-alpha.351
+
 ## 0.2.0-alpha.350
 
 ### Patch Changes