+++ plugins/unified-inbox/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @brains/unified-inbox
 
+## 0.2.0-alpha.351
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/ui-library@0.2.0-alpha.351
+  - @brains/utils@0.2.0-alpha.351
+  - @brains/plugins@0.2.0-alpha.351
+
 ## 0.2.0-alpha.350
 
 ### Patch Changes