+++ shared/app-ui-react/CHANGELOG.md
@@ -1,3 +1,5 @@
 # @brains/app-ui-react
 
+## 0.2.0-alpha.351
+
 ## 0.2.0-alpha.350