+++ shared/deploy-support/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brains/deploy-support
 
+## 0.2.0-alpha.351
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/utils@0.2.0-alpha.351
+
 ## 0.2.0-alpha.350
 
 ### Patch Changes