+++ shell/messaging-service/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @brains/messaging-service
 
+## 0.2.0-alpha.351
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.351
+  - @brains/operation-context@0.2.0-alpha.351
+  - @brains/utils@0.2.0-alpha.351
+  - @brains/templates@0.2.0-alpha.351
+
 ## 0.2.0-alpha.350
 
 ### Patch Changes