+++ shell/templates/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @brains/templates
 
+## 0.2.0-alpha.351
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/content-formatters@0.2.0-alpha.351
+  - @brains/contracts@0.2.0-alpha.351
+  - @brains/utils@0.2.0-alpha.351
+
 ## 0.2.0-alpha.350
 
 ### Patch Changes